Added Epic Free Games Section and Search Box.
Made a small section for Epic's Free games for the front page. Also added a search box to the nav bar.
The search box takes in a title and produces results in a dropdown. The results are links to the game page.
Made numerous css changes and reorganized some components.